index benchmark update
